# api/openapi.yaml
openapi: 3.0.3
info:
  title: Trading Platform API
  description: |
    Public entry is the API gateway on :8080. Behind it: auth-service, user-service, market-data-service.
    Error body is always `{ "error": { "code", "message", "details?" } }`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: API gateway (docker-compose public entry)
...
paths:
...
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Current user profile
      operationId: getMe
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Users]
      summary: Update current user display name
      operationId: patchMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchMeRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation or malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                - field: email
                  message: email is required
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Missing or invalid access token
...
  schemas:
...
    PatchMeRequest:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          maxLength: 100
          example: New Name
    User:
      type: object
      required: [id, email, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - BAD_GATEWAY
                - UNAVAILABLE
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
security:
  - bearerAuth: []